requirements.txt
openai
httpx
python-dotenv
pydantic
